+++ demo/RadioField/RadioFieldDemo.jsx
@@ -6,13 +6,12 @@
  * All rights reserved.
  */
 
+import React from 'react';
 import Group from 'salt-group';
-
 import RadioField from 'salt-radio-field';
 
 class Demo extends React.Component {
   render() {
-    const t = this;
     const radioFieldProps1 = {
       data: [
         {
@@ -25,14 +24,14 @@ class Demo extends React.Component {
             <table className="demoTable">
               <tbody>
                 <tr>
-                  <td className="avatar-td"><img className="avatar" src="https://img.alicdn.com/tps/TB1vZnyJFXXXXX5XpXXXXXXXXXX-32-32.png" /></td>
+                  <td className="avatar-td"><img alt="" className="avatar" src="https://img.alicdn.com/tps/TB1vZnyJFXXXXX5XpXXXXXXXXXX-32-32.png" /></td>
                   <td className="info-td"><div className="name">周姮</div><div className="postName">资深交互设计师</div></td>
                 </tr>
               </tbody>
             </table>
           ),
           disable: false,
-          label: '周姮'
+          label: '周姮',
         },
         {
           value: {
@@ -44,14 +43,14 @@ class Demo extends React.Component {
             <table className="demoTable">
               <tbody>
                 <tr>
-                  <td className="avatar-td"><img className="avatar" src="https://img.alicdn.com/tps/TB1CmDsJFXXXXcxXpXXXXXXXXXX-32-32.png" /></td>
+                  <td className="avatar-td"><img alt="" className="avatar" src="https://img.alicdn.com/tps/TB1CmDsJFXXXXcxXpXXXXXXXXXX-32-32.png" /></td>
                   <td className="info-td"><div className="name">李伟（孟则）</div><div className="postName">资深交互设计师</div></td>
                 </tr>
               </tbody>
             </table>
           ),
           disable: false,
-          label: '李伟（孟则）'
+          label: '李伟（孟则）',
         },
         {
           value: 'plain',
@@ -79,7 +78,7 @@ class Demo extends React.Component {
             <table className="demoTable">
               <tbody>
                 <tr>
-                  <td className="avatar-td"><img className="avatar" src="https://img.alicdn.com/tps/TB1vZnyJFXXXXX5XpXXXXXXXXXX-32-32.png" /></td>
+                  <td className="avatar-td"><img alt="" className="avatar" src="https://img.alicdn.com/tps/TB1vZnyJFXXXXX5XpXXXXXXXXXX-32-32.png" /></td>
                   <td className="info-td"><div className="name">周姮</div><div className="postName">资深交互设计师</div></td>
                 </tr>
               </tbody>
@@ -97,7 +96,7 @@ class Demo extends React.Component {
             <table className="demoTable">
               <tbody>
                 <tr>
-                  <td className="avatar-td"><img className="avatar" src="https://img.alicdn.com/tps/TB1CmDsJFXXXXcxXpXXXXXXXXXX-32-32.png" /></td>
+                  <td className="avatar-td"><img alt="" className="avatar" src="https://img.alicdn.com/tps/TB1CmDsJFXXXXcxXpXXXXXXXXXX-32-32.png" /></td>
                   <td className="info-td"><div className="name">李伟（孟则）</div><div className="postName">资深交互设计师</div></td>
                 </tr>
               </tbody>
@@ -131,7 +130,7 @@ class Demo extends React.Component {
         <Group>
           <Group.Head className="t-FS14 t-LH1_5 t-LH20 t-PT10 t-PB10 t-PL18">popup 显示</Group.Head>
           <Group.List>
-            <RadioField {...radioFieldProps1} layoutType="popup" />
+            <RadioField {...radioFieldProps1} mode="popup" label="复选" />
             <RadioField {...radioFieldProps2} iconPosition="right" label="单选（图标在右）" />
           </Group.List>
         </Group>